sounds and forge

+++ code/modules/mining/oresningots.dm
@@ -337,18 +337,23 @@
 	density = 1
 	anchored = 1
 	var/coalfed = 0
+	var/ironingot = 0
 
 /obj/structure/forge/attackby(var/obj/item/newore/O, var/mob/living/carbon/human/user)
+	if (!(istype(O,	/obj/item/newore)))
+		visible_message("That isn't a valid ore!")
+		return
 	if (istype(O,	/obj/item/newore/coalore))
 		visible_message("[user] feeds coal into the fire!")
 		src.coalfed = 1 //lets not heat everyones forge
+		playsound(src, 'sound/effects/adding_coal.ogg', 100, 1, 1)
 		return
-
 	if(coalfed == 0)
 		to_chat(src, "The forge requires coal to burn hot enough to smelt ore!")
 		return
 	else if (istype(O,	/obj/item/newore/copperore))
 		visible_message("[user] feeds [O] into the fire!")
+		playsound(src, 'sound/effects/furnace_fire.ogg', 100, 1, 1)
 		sleep(60) //10 seconds
 		new /obj/item/copperingot(src)
 		qdel(O)
@@ -356,6 +361,7 @@
 		return
 	else if (istype(O,	/obj/item/newore/silverore))
 		visible_message("[user] feeds [O] into the fire!")
+		playsound(src, 'sound/effects/furnace_fire.ogg', 100, 1, 1)
 		sleep(60) //10 seconds
 		new /obj/item/silveringot(src)
 		qdel(O)
@@ -363,6 +369,7 @@
 		return
 	else if (istype(O,	/obj/item/newore/goldore))
 		visible_message("[user] feeds [O] into the fire!")
+		playsound(src, 'sound/effects/furnace_fire.ogg', 100, 1, 1)
 		sleep(60) //10 seconds
 		new /obj/item/goldingot(src)
 		qdel(O)
@@ -370,10 +377,18 @@
 		return
 	else if (istype(O,	/obj/item/newore/ironore))
 		visible_message("[user] feeds [O] into the fire!")
+		playsound(src, 'sound/effects/furnace_fire.ogg', 100, 1, 1)
 		sleep(60) //10 seconds
 		new /obj/item/ironingot(src)
 		qdel(O)
 		src.coalfed = 0
 		return
-
+	else if (istype(O,	/obj/item/newore/adamantiumore))
+		visible_message("[user] feeds [O] into the fire!")
+		playsound(src, 'sound/effects/furnace_fire.ogg', 100, 1, 1)
+		sleep(60) //10 seconds
+		new /obj/item/adamantiumingot(src)
+		qdel(O)
+		src.coalfed = 0
+		return
